refactor auth form
