+++ src/components/signupform/BaseSignupForm.js
@@ -7,6 +7,52 @@ import Link from "@material-ui/core/Link";
 import RouterLink from "../RouterLink";
 import Grid from "@material-ui/core/Grid";
 import { makeStyles } from "@material-ui/core/styles";
+import { string, ref } from "yup";
+
+
+export const baseInitialValues = {
+  firstName: "",
+  lastName: "",
+  email: "",
+  password: "",
+  verifyPassword: ""
+};
+
+export const baseValidationFields = {
+  firstName: string()
+    .max(15, "Must be 15 characters or less")
+    .required("Required"),
+  lastName: string()
+    .max(20, "Must be 20 characters or less")
+    .required("Required"),
+  email: string()
+    .email("Invalid email addresss`")
+    .required("Required"),
+  password: string()
+    .min(8, "Must be 8 characters or more")
+    .max(40, "Must be 40 characters or less")
+    .matches(
+      /^(?=.*?[#?!@$%^&*-]).{8,}$|.{15}/,
+      "Must have at least one special character or be at least 15 characters long"
+    )
+    .matches(
+      /^(?=.*?[A-Z]).{8,}$|.{15}/,
+      "Must have at least one uppercase character or be at least 15 characters long"
+    )
+    .matches(
+      /^(?=.*?[a-z]).{8,}$|.{15}/,
+      "Must have at least one lowercase character or be at least 15 characters long"
+    )
+    .matches(
+      /^(?=.*?[0-9]).{8,}$|.{15}/,
+      "Must have at least one integer or be at least than 15 characters long"
+    )
+    .required("Required"),
+  verifyPassword: string().oneOf(
+    [ref("password"), null],
+    "Passwords must match"
+  )
+};
 
 const useStyles = makeStyles(theme => ({
   form: {
@@ -18,6 +64,13 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
+const fieldAttributes = {
+  variant: "outlined",
+  required: true,
+  component: TextField,
+  fullWidth: true
+};
+
 export default function BaseSignupForm({ component: Component, isSubmitting }) {
   const classes = useStyles();
   return (
@@ -27,25 +80,19 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
           <Field
             autoComplete="fname"
             name="firstName"
-            variant="outlined"
-            required
-            component={TextField}
-            fullWidth
             id="firstName"
             label="First Name"
             autoFocus
+            {...fieldAttributes}
           />
         </Grid>
         <Grid item xs={12} sm={6}>
           <Field
-            variant="outlined"
-            required
-            component={TextField}
-            fullWidth
             id="lastName"
             label="Last Name"
             name="lastName"
             autoComplete="lname"
+            {...fieldAttributes}
           />
         </Grid>
         <Grid item xs={12} sm={6}>
@@ -56,21 +103,44 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
         </Grid>
         <Grid item xs={12}>
           <Field
-            variant="outlined"
-            required
-            component={TextField}
-            fullWidth
             id="email"
             label="Email Address"
             type="email"
             name="email"
             autoComplete="email"
-          />
+            {...fieldAttributes}
+            />
         </Grid>
         <Grid item xs={12}>
           <ErrorMessage name="email" component="div" />
         </Grid>
         {Component && <Component />}
+        <Grid item xs={12}>
+          <Field
+            id="password"
+            label="Password"
+            type="password"
+            name="password"
+            autoComplete="email"
+            {...fieldAttributes}
+          />
+        </Grid>
+        <Grid item xs={12}>
+          <ErrorMessage name="password" component="div" />
+        </Grid>
+        <Grid item xs={12}>
+          <Field
+            id="verifyPassword"
+            label="Verify Password"
+            type="password"
+            name="verifyPassword"
+            autoComplete="email"
+            {...fieldAttributes}
+          />
+        </Grid>
+        <Grid item xs={12}>
+          <ErrorMessage name="verifyPassword" component="div" />
+        </Grid>
       </Grid>
       <Button
         type="submit"
