+++ src/components/signupform/DonorFields.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
 import { Field, ErrorMessage } from "formik";
-import { string, ref } from "yup";
+import { string } from "yup";
 
 import TextField from "@material-ui/core/TextField";
 import Grid from "@material-ui/core/Grid";
@@ -15,8 +15,6 @@ export const donorInitialValues = {
   street: '',
   lg: '',
   state: '',
-  password: '',
-  verifyPassword: '',
 }
 
 export const donorValidationFields = {
@@ -27,31 +25,7 @@ export const donorValidationFields = {
     .max(40, "Must be 40 characters or less")
     .required("Required"),
   lg: location,
-  state: location,
-  password: string()
-    .min(8, "Must be 8 characters or more")
-    .max(40, "Must be 40 characters or less")
-    .matches(
-      /^(?=.*?[#?!@$%^&*-]).{8,}$|.{15}/,
-      "Must have at least one special character or be at least 15 characters long"
-    )
-    .matches(
-      /^(?=.*?[A-Z]).{8,}$|.{15}/,
-      "Must have at least one uppercase character or be at least 15 characters long"
-    )
-    .matches(
-      /^(?=.*?[a-z]).{8,}$|.{15}/,
-      "Must have at least one lowercase character or be at least 15 characters long"
-    )
-    .matches(
-      /^(?=.*?[0-9]).{8,}$|.{15}/,
-      "Must have at least one integer or be at least than 15 characters long"
-    )
-    .required("Required"),
-  verifyPassword: string().oneOf(
-    [ref("password"), null],
-    "Passwords must match"
-  )
+  state: location
 };
 
 export default function DonorFields({ handleChange }) {
@@ -79,20 +53,6 @@ export default function DonorFields({ handleChange }) {
       label: "State",
       id: "state",
       autoComplete: "state"
-    },
-    {
-      name: "password",
-      label: "Password",
-      id: "password",
-      type: "password",
-      autoComplete: "password"
-    },
-    {
-      name: "verifyPassword",
-      label: "Verify Password",
-      id: "verifyPassword",
-      type: "password",
-      autoComplete: "verifyPassword"
     }
   ];
   return (
