+++ src/components/signupform/DonorSignup.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { object } from "yup";
-import PartialForm, {baseInitialValues, baseValidationFields} from "./signupform/PartialForm";
+import AuthForm from "./signupform/AuthForm";
+import BaseSignupForm, {baseInitialValues, baseValidationFields} from "./signupform/BaseSignupForm";
 import DonorFields, {donorInitialValues, donorValidationFields} from "./signupform/DonorFields";
 import { signup } from "../../actions/auth";
 
@@ -9,11 +10,12 @@ export const initialValues = {...baseInitialValues, ...donorInitialValues};
 export const validationSchema = object({...baseValidationFields, ...donorValidationFields});
 
 export default () => (
-  <PartialForm
+  <AuthForm
     action={signup}
-    component={DonorFields}
     redirectPath="/dashboard"
     validationSchema={validationSchema}
     initialValues={initialValues}
-  />
+  >
+  <BaseSignupForm component={DonorFields} />
+</AuthForm>
 );
